Initial Commit
Adds support for the NanoPi 3 and uses WIC to create a bootable SD image

FriendlyArm documentation shows that building the images for all of the
s5p6818 and s5p4418 boards have the same git trees and config files.
Duplicate the NanoPi Fire3 config for the other boards and tweak their
names.

Signed-off-by: Jon Mason <jdmason@kudzu.us>